Move selectors to their just and rightful place
