+++ client-v2/src/util/form.ts
@@ -2,16 +2,6 @@ import { ArticleFragmentMeta } from 'shared/types/Collection';
 import { DerivedArticle } from 'shared/types/Article';
 import omit from 'lodash/omit';
 import compact from 'lodash/compact';
-import { selectors } from 'shared/bundles/collectionsBundle';
-import {
-  selectSharedState,
-  createArticleFromArticleFragmentSelector
-} from 'shared/selectors/shared';
-import { getCollectionConfig } from 'selectors/frontsSelectors';
-import { hasMainVideo } from 'shared/util/derivedArticle';
-import { isCollectionConfigDynamic } from './frontsUtils';
-import { createSelector } from 'reselect';
-import { State } from 'types/State';
 
 export interface ArticleFragmentFormData {
   headline: string;
@@ -38,80 +28,6 @@ interface ImageData {
   origin?: string;
   thumb?: string;
 }
-
-export const defaultFields = [
-  'headline',
-  'isBoosted',
-  'showQuotedHeadline',
-  'showBoostedHeadline',
-  'customKicker',
-  'isBreaking',
-  'byline',
-  'showByline',
-  'trailText',
-  'imageCutoutReplace',
-  'imageHide',
-  'imageSlideshowReplace',
-  'primaryImage',
-  'cutoutImage',
-  'slideshow'
-];
-
-export const supportingFields = [
-  'headline',
-  'customKicker',
-  'isBreaking',
-  'showKickerSection',
-  'showKickerCustom'
-];
-
-const selectIsSupporting = (
-  _: unknown,
-  __: unknown,
-  isSupporting: boolean
-) => isSupporting;
-
-const selectParentCollectionConfig = (
-  state: State,
-  articleFragmentId: string
-) => {
-  const collectionId = selectors.selectParentCollectionOfArticleFragment(
-    selectSharedState(state),
-    articleFragmentId
-  );
-  return collectionId ? getCollectionConfig(state, collectionId) : undefined;
-};
-
-export const createSelectFormFieldsForCollectionItem = () => {
-  const selectDerivedArticle = createArticleFromArticleFragmentSelector();
-  const selectDerivedArticleFromRootState = (state: State, id: string) =>
-    selectDerivedArticle(selectSharedState(state), id);
-  return createSelector(
-    selectDerivedArticleFromRootState,
-    selectParentCollectionConfig,
-    selectIsSupporting,
-    (derivedArticle, parentCollectionConfig, isSupporting) => {
-      if (!derivedArticle) {
-        return [];
-      }
-      if (isSupporting) {
-        return supportingFields;
-      }
-      const fields = defaultFields.slice();
-
-      if (isCollectionConfigDynamic(parentCollectionConfig)) {
-        fields.push('isBoosted');
-      }
-      if (derivedArticle.liveBloggingNow === 'true') {
-        fields.push('showLivePlayable');
-      }
-      if (hasMainVideo(derivedArticle)) {
-        fields.push('showMainVideo');
-      }
-      return fields;
-    }
-  );
-};
 
 const strToInt = (str: string | void) => (str ? parseInt(str, 10) : undefined);
 const intToStr = (int: number | void) => (int ? int.toString() : undefined);
